Adding fields to generated entity classes.
https://trello.com/c/JcNP0RMb

We use Javassist to add fields to the new class definitions. We also
add appropriate initializers based on the default value. This commit also
restructures the builder classes and adds methods which makes us ready for
bundle generations.

What is not yet handled:

* jdo metadata for fields
* full transition to using a mds entities bundle

Support for regenerating classes in our classloaders was added, however
we still face a problem of re-registering jdo metada. This should be
resolved with transitioning to a bundle.

Change-Id: Idfb268804f9b0f11d68dac38e0cef0e7634b281a